+++ src/main/java/step2/move/MoveStrategy.java
@@ -1,4 +1,4 @@
-package step2.Model;
+package step2.move;
 
 public interface MoveStrategy {
     boolean isPossibleToGo();
